+++ src/main/java/janggi/domain/piece/Ma.java
@@ -1,17 +1,19 @@
-package janggi.domain;
+package janggi.domain.piece;
 
 import static java.lang.Math.abs;
 
+import janggi.domain.Point;
+import janggi.domain.status.Team;
 import java.util.List;
 
-public class Sang implements Piece {
+public class Ma implements Piece {
 
-    private static final int DISTANCE_MAX = 3;
-    private static final int DISTANCE_MIN = 2;
+    private static final int DISTANCE_MAX = 2;
+    private static final int DISTANCE_MIN = 1;
 
     private final Team team;
 
-    public Sang(Team team) {
+    public Ma(Team team) {
         this.team = team;
     }
 
@@ -26,22 +28,15 @@ public class Sang implements Piece {
         int pathY = to.getY() - from.getY();
         int distanceX = abs(pathX);
         int distanceY = abs(pathY);
-        int signX = Integer.compare(pathX, 0);
-        int signY = Integer.compare(pathY, 0);
 
         if (!((distanceX == DISTANCE_MAX && distanceY == DISTANCE_MIN) ||
                 (distanceX == DISTANCE_MIN && distanceY == DISTANCE_MAX))) {
             throw new IllegalArgumentException();
         }
         if (distanceX == DISTANCE_MAX) {
-            return List.of(
-                    Point.of(from.getX() + (pathX / DISTANCE_MAX), from.getY()),
-                    Point.of(from.getX() + signX * DISTANCE_MIN,  from.getY() + signY)
-            );
+            return List.of(Point.of(from.getX() + (pathX / DISTANCE_MAX), from.getY()));
         }
-        return List.of(
-                Point.of(from.getX(), from.getY() + (pathY / DISTANCE_MAX)),
-                Point.of(from.getX() + signX,  from.getY() + signY * DISTANCE_MIN));
+        return List.of(Point.of(from.getX(), from.getY() + (pathY / DISTANCE_MAX)));
     }
 
     @Override
