+++ src/main/java/janggi/domain/piece/Pho.java
@@ -1,5 +1,7 @@
-package janggi.domain;
+package janggi.domain.piece;
 
+import janggi.domain.Point;
+import janggi.domain.status.Team;
 import java.util.List;
 
 public class Pho implements Piece {
