+++ src/main/java/janggi/domain/status/GameStatus.java
@@ -1,4 +1,6 @@
-package janggi.domain;
+package janggi.domain.status;
+
+import janggi.domain.Point;
 
 public interface GameStatus {
     GameStatus move(Point from, Point to);
