+++ src/main/java/janggi/domain/status/HanTurn.java
@@ -1,4 +1,6 @@
-package janggi.domain;
+package janggi.domain.status;
+
+import janggi.domain.Point;
 
 public class HanTurn implements GameStatus {
